GUI.py: Fixed glitch that said status OK when you couldn't connect to server
If you have wrong username/password or can't connect in some way you got a
green OK between the ERROR/ERR flashes.

Fixed so the systray icon should be black, and the floating becomes blue
with white text and say ERR/ERROR, depending if you use LONG or not.

this fixes SVN ID: 3020462

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>