servers: Start to remove LXML
Apparently it's an annoying dependency when making nagstamon
cross-platform.

Replace it with BeatifulSoup. BeautifulSoup is pure-python (one single
BSD-licensed file that could easily be distributed with nagstamon), and
can handle most kinds of invalid HTML (as well as valid XML) very well.

This patch introduces a LxmlFreeGenericServer class, that simply
overrides the LXML bits in the normal GenericServer. My thinking is
that, at some point, all the servers will be ported to use this class,
at which point the LXML dependency will be no more and this class can
replace the current GenericServer.

Ninja and Nagios has been ported to this class. I have done some
rudimentary testing to see that on my server they show me the same data
as they did before, but this is admittedly very shallow testing, and
there are probably several regressions. Think of it as a proof of
concept.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>